refactor: move getAge function to helpers
